+++ scout/commands/update/base.py
@@ -2,9 +2,6 @@ import logging
 
 import click
 
-from scout.update.panel import update_panel
-from scout.utils.date import get_date
-
 from .case import case as case_command
 from .compounds import compounds as compound_command
 from .disease import diseases as disease_command
@@ -14,6 +11,7 @@ from .individual import individual as individual_command
 from .institute import institute as institute_command
 from .omim import omim as omim_command
 from .panel import panel as panel_command
+from .panelapp import panelapp_green as panelapp_green_command
 from .phenotype_groups import groups as groups_command
 from .user import user as user_command
 
@@ -34,6 +32,7 @@ update.add_command(user_command)
 update.add_command(panel_command)
 update.add_command(case_command)
 update.add_command(omim_command)
+update.add_command(panelapp_green_command)
 update.add_command(compound_command)
 update.add_command(hpo_command)
 update.add_command(gene_command)