+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
+## [unreleased]
+### Added
+- Create/Update a gene panel contaning all PanelApp green genes (`scout update panelapp-green -i <cust_id>`)
 
 ## [4.63]
 ### Added